+++ package.json
@@ -72,7 +72,7 @@
     "puppeteer-core": "^19.9.0",
     "sharp": "^0.32.1",
     "ts-jest": "^28.0.8",
-    "typescript": "^5.1.3",
+    "typescript": "5.1.3",
     "yarn": "^1.22.19",
     "langchain":"0.0.125"
   },
