+++ plugin-src/summarizeFile.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from "fs";
 
 
-import { Box, log, FileBox, MenuItemFile, RenderElements, contextMenu, } from "../src/pluginFacade"
+import { Box, log, FileBox, MenuItemFile, RenderElements, contextMenu, } from "../dist/pluginFacade"
 
 import { SummarizerFactory } from "./summarizeFile/summarizer-factory";
 import { openaiSummaryLLMFactory } from "./summarizeFile/llms/summary-openai";
@@ -10,8 +10,8 @@ import { ConfigManager } from "./summarizeFile/config-manager";
 const SUMMARY_FIELD_NAME = 'summary';
 
 
-contextMenu.addFileBoxMenuItem((item) => new MenuItemFile({ label: "Summarize File", click: () => console.error("hi")  }));
-// summarize(item,new SummarizerFactory(openaiSummaryLLMFactory,new ConfigManager()))
+contextMenu.addFileBoxMenuItem((item) => new MenuItemFile({ label: "Summarize File", click: () => 
+                     summarize(item,new SummarizerFactory(openaiSummaryLLMFactory,new ConfigManager()))}));
 
 Box.Tabs.register({
     name: 'Summary',
