--- plugin-src/summarizeFile/config-manager.test.ts
@@ -1,8 +1,0 @@
-describe("ConfigManager", () => {
-    it('loads apiKey form the config if the config exists',()=>{
-        //TODO
-    });
-    it('asks for an api key and saves it if the config does not exist',()=>{
-        //TODO
-    });
-});
