+++ plugin-src/tsconfig.json
@@ -8,7 +8,6 @@
     "noImplicitAny": false,
     "noImplicitOverride": true,
     "outDir": "../plugin",
-    "rootDir": ".",
     "moduleResolution": "node"
   },
   "exclude": ["**/*.test.ts"]
